# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tracking calorie intake and expenditure",
        "title": "Calorie Deficit API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/daily-intake": {
            "get": {
                "description": "Get a list of daily intakes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "daily-intake"
                ],
                "summary": "Get daily intakes list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Date in YYYY-MM-DD format",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Meal type (breakfast, brunch, lunch, afternoonSnack, dinner, lateNightSnack)",
                        "name": "meal_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dailyintakedto.DailyIntakeListResponseSuccessDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.AppError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new daily intake record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "daily-intake"
                ],
                "summary": "Create a daily intake",
                "parameters": [
                    {
                        "description": "Daily Intake Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dailyintake.DailyIntakeCreateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dailyintake.DailyIntakeCreateResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.AppError"
                        }
                    }
                }
            }
        },
        "/daily-intake/{id}": {
            "get": {
                "description": "Get a daily intake by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "daily-intake"
                ],
                "summary": "Get a daily intake",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Daily Intake ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dailyintake.DailyIntakeCreateResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.AppError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update an existing daily intake record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "daily-intake"
                ],
                "summary": "Update a daily intake",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Daily Intake ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Daily Intake Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dailyintake.DailyIntakeUpdateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dailyintake.DailyIntakeCreateResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/types.AppError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/types.AppError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "constants.MealType": {
            "type": "string",
            "enum": [
                "breakfast",
                "brunch",
                "lunch",
                "afternoonSnack",
                "dinner",
                "lateNightSnack",
                "drinks",
                "other"
            ],
            "x-enum-comments": {
                "Drinks": "definitely designed for Taiwanese people"
            },
            "x-enum-varnames": [
                "Breakfast",
                "Brunch",
                "Lunch",
                "AfternoonSnack",
                "Dinner",
                "LateNightSnack",
                "Drinks",
                "Other"
            ]
        },
        "constants.MeasurementType": {
            "type": "string",
            "enum": [
                "grams",
                "ml"
            ],
            "x-enum-varnames": [
                "Grams",
                "ML"
            ]
        },
        "dailyintake.DailyIntakeCreateRequestDTO": {
            "type": "object",
            "required": [
                "date",
                "meal_items",
                "meal_type",
                "user_id"
            ],
            "properties": {
                "date": {
                    "description": "Date in ISO 8601",
                    "type": "string"
                },
                "meal_items": {
                    "description": "MealItems is a slice of MealItem",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dailyintakedto.MealItemRequestDTO"
                    }
                },
                "meal_type": {
                    "description": "MealType is an enum for the meal type",
                    "enum": [
                        "breakfast",
                        "lunch",
                        "dinner",
                        "snack"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/constants.MealType"
                        }
                    ]
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "dailyintake.DailyIntakeCreateResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt in ISO 8601 format 2025-05-18T02:55:46.844Z",
                    "type": "string"
                },
                "date": {
                    "description": "Date in ISO 8601 format 2025-05-18T02:55:46.844Z",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the primary key of the daily intake table",
                    "type": "integer"
                },
                "meal_items": {
                    "description": "MealItems is a slice of MealItem",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dailyintakedto.MealItemResponseDTO"
                    }
                },
                "meal_type": {
                    "description": "MealType is an enum for the meal type",
                    "allOf": [
                        {
                            "$ref": "#/definitions/constants.MealType"
                        }
                    ]
                },
                "updated_at": {
                    "description": "UpdatedAt in ISO 8601 format 2025-05-18T02:55:46.844Z",
                    "type": "string"
                },
                "user_id": {
                    "description": "UserID is the foreign key to the user table",
                    "type": "integer"
                }
            }
        },
        "dailyintake.DailyIntakeUpdateRequestDTO": {
            "type": "object",
            "required": [
                "meal_items"
            ],
            "properties": {
                "meal_items": {
                    "description": "UserID    uint                 `json:\"user_id\"`                                             // UserID is the foreign key to the user table\nDate      string               `json:\"date\" binding:\"datetime=2006-01-02T15:04:05.000Z\"`       // Date in ISO 8601\nMealType  constants.MealType   `json:\"meal_type\" binding:\"oneof=breakfast lunch dinner snack\"` // MealType is an enum for the meal type",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dailyintakedto.MealItemResponseDTO"
                    }
                }
            }
        },
        "dailyintakedto.DailyIntakeCreateResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "CreatedAt in ISO 8601 format 2025-05-18T02:55:46.844Z",
                    "type": "string"
                },
                "date": {
                    "description": "Date in ISO 8601 format 2025-05-18T02:55:46.844Z",
                    "type": "string"
                },
                "id": {
                    "description": "ID is the primary key of the daily intake table",
                    "type": "integer"
                },
                "meal_items": {
                    "description": "MealItems is a slice of MealItem",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dailyintakedto.MealItemResponseDTO"
                    }
                },
                "meal_type": {
                    "description": "MealType is an enum for the meal type",
                    "allOf": [
                        {
                            "$ref": "#/definitions/constants.MealType"
                        }
                    ]
                },
                "updated_at": {
                    "description": "UpdatedAt in ISO 8601 format 2025-05-18T02:55:46.844Z",
                    "type": "string"
                },
                "user_id": {
                    "description": "UserID is the foreign key to the user table",
                    "type": "integer"
                }
            }
        },
        "dailyintakedto.DailyIntakeListResponseDTO": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "Items is a slice of DailyIntakeCreateResponseDTO",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dailyintakedto.DailyIntakeCreateResponseDTO"
                    }
                },
                "meta": {
                    "description": "Meta is a pointer to the Meta struct",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Meta"
                        }
                    ]
                }
            }
        },
        "dailyintakedto.DailyIntakeListResponseSuccessDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dailyintakedto.DailyIntakeListResponseDTO"
                    }
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "description": "Pointer to Meta struct",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.Meta"
                        }
                    ]
                }
            }
        },
        "dailyintakedto.MealItemRequestDTO": {
            "type": "object",
            "required": [
                "measurement",
                "name",
                "quantity"
            ],
            "properties": {
                "calorie": {
                    "description": "Calorie of the meal item",
                    "type": "number"
                },
                "measurement": {
                    "description": "Measurement is an enum for the measurement type",
                    "enum": [
                        "grams",
                        "ml"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/constants.MeasurementType"
                        }
                    ]
                },
                "name": {
                    "description": "Name of the meal item",
                    "type": "string"
                },
                "quantity": {
                    "description": "Quantity of the meal item",
                    "type": "number"
                }
            }
        },
        "dailyintakedto.MealItemResponseDTO": {
            "type": "object",
            "properties": {
                "calorie": {
                    "description": "Calorie of the meal item",
                    "type": "number"
                },
                "daily_intake_id": {
                    "description": "DailyIntakeID is the foreign key to the daily intake table",
                    "type": "integer"
                },
                "id": {
                    "description": "ID is the primary key of the meal item table",
                    "type": "integer"
                },
                "measurement": {
                    "description": "Measurement is an enum for the measurement type",
                    "allOf": [
                        {
                            "$ref": "#/definitions/constants.MeasurementType"
                        }
                    ]
                },
                "name": {
                    "description": "Name of the meal item",
                    "type": "string"
                },
                "quantity": {
                    "description": "Quantity of the meal item",
                    "type": "number"
                }
            }
        },
        "types.AppError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "types.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                }
            }
        }
    }
}